+++ apps/react/src/index.ts
@@ -1,0 +1,6 @@
+import C1Input from './components/C1Input';
+import C1LintFixture from './components/C1LintFixture';
+
+export { C1Input, C1LintFixture };
+
+export default C1LintFixture;