+++ apps/react/vite.config.ts
@@ -1,0 +1,20 @@
+import { resolve } from 'node:path';
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    lib: {
+      entry: resolve(__dirname, 'src/index.ts'),
+      name: 'C1ReactLibrary',
+      formats: ['es'],
+      fileName: 'index',
+      cssFileName: 'style',
+    },
+    rollupOptions: {
+      external: ['react', 'react-dom'],
+    },
+    outDir: 'dist',
+  },
+});